Botaneer patch 1
Started a new game project about growing and creativity. 
Added a simple grid showing pot area. This pot area was further expanded in a mode that would allow creating/editing pots in a future update. This mode can be accessed by pressing "TAB". Basic controls are WASD to move around the grid and LEFT/RIGHT arrows to zoom in/out. In a future update, This mode would have a tool sidebar to select zoom, build, break and color options for the pot design mode.