Botaneer patch 10
Created recursive algorithm for determining connectivity. All that's left is to sort out the scenario's when a split path occurs (2 function calls instead of 1). After, the one with the highest y is taken into consideration (per side) with the lowest one when taking into account both sides of the pot.

This update resolves #14 where this function checks for connectivity, which considers a valid base with sides that are connected into the base. Issue #14 describes the previous update that didn't consider connectivity and included pot pieces not connected to the base (or the pot of interest).