First version of the Interactive Code Map
You can click to place a point and hold a point to drag it to a new location. Currently, clicking a point shows it's menu but there's no functionality yet. Future implementations will add name tags to show what the point is describing, a description booklet to type paragraphs about that point, connecting points with a visible line, and other visual changes (point color and line color to name a few).